removed double scrollbar | refractored code | put the loader in middle
