+++ components/incidents/styles/styles.module.css
@@ -13,6 +13,7 @@
 }
 
 .spinner {
+    margin-top : 35vh;
     display: flex;
     justify-content: center;
     align-items: center;
